# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  swagmodel.GetDraft:
    properties:
      body:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
      imageURL:
        type: string
      title:
        type: string
    type: object
  swagmodel.GetInbox:
    properties:
      body:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
      receiverID:
        type: integer
    type: object
  swagmodel.GetMedia:
    properties:
      alt:
        type: string
      caption:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
      url:
        type: string
    type: object
...
  swagmodel.InputDraft:
    properties:
      body:
        type: string
      creatorID:
        type: integer
      id:
        type: integer
      imageURL:
        type: string
      title:
        type: string
    type: object
  swagmodel.InputInbox:
    properties:
      body:
        type: string
    type: object
  swagmodel.InputMedia:
    properties:
      alt:
        type: string
      caption:
        type: string
      url:
        type: string
    type: object
...
  swagmodel.Response:
    properties:
      msg:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: hdkef11@gmail.com
    name: Hadekha Erfadila Fitra
  description: Bloggo is a REST Api for blogging
  title: Bloggo API
  version: "1.0"
paths:
...
  /drafts:
    get:
      description: get all drafts navigated by last-id and limit.
      parameters:
      - description: ID of the last draft in recent array of draft
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many draft you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetDraft'
            type: array
      summary: get all drafts
      tags:
      - drafts
    post:
      consumes:
      - application/json
      description: create an draft
      parameters:
      - description: create an draft
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputDraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an draft
      tags:
      - drafts
  /drafts/{id}:
    delete:
      consumes:
      - application/json
      description: delete an draft
      parameters:
      - description: ID of draft
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an draft
      tags:
      - drafts
    get:
      description: get detail of draft specified by id
      parameters:
      - description: ID of draft
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.GetDraft'
      summary: get one draft
      tags:
      - drafts
    put:
      consumes:
      - application/json
      description: update an draft
      parameters:
      - description: ID of draft
        in: path
        name: id
        required: true
        type: integer
      - description: update an draft
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputDraft'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an draft
      tags:
      - drafts
  /inboxes/{id}:
    delete:
      consumes:
      - application/json
      description: delete an inbox
      parameters:
      - description: ID of inbox
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an inbox
      tags:
      - inboxes
    put:
      consumes:
      - application/json
      description: update an inbox
      parameters:
      - description: ID of inbox
        in: path
        name: id
        required: true
        type: integer
      - description: update an inbox
        in: body
        name: inbox
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputInbox'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an inbox
      tags:
      - inboxes
  /inboxes/{receiver-id}:
    get:
      description: get all inboxes navigated by last-id and limit.
      parameters:
      - description: ID of user / receiver
        in: path
        name: receiver-id
        required: true
        type: integer
      - description: ID of the last inbox in recent array of inboxes
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many inboxes you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetInbox'
            type: array
      summary: get all inboxes
      tags:
      - inboxes
    post:
      consumes:
      - application/json
      description: create an inbox
      parameters:
      - description: ID of user / receiver
        in: path
        name: receiver-id
        required: true
        type: integer
      - description: create an inbox
        in: body
        name: inbox
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputInbox'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an inbox
      tags:
      - inboxes
...
  /media:
    get:
      description: get all media navigated by last-id and limit.
      parameters:
      - description: ID of the last media in recent array of media
        in: query
        name: last-id
        required: true
        type: integer
      - description: how many media you want to take
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/swagmodel.GetMedia'
            type: array
      security:
      - AuthToken: []
      summary: get all media
      tags:
      - media
    post:
      consumes:
      - application/json
      description: create an media
      parameters:
      - description: create an media
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: create an media
      tags:
      - media
  /media/{id}:
    delete:
      consumes:
      - application/json
      description: delete an media
      parameters:
      - description: ID of media
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: delete an media
      tags:
      - media
    put:
      consumes:
      - application/json
      description: update an media
      parameters:
      - description: ID of media
        in: path
        name: id
        required: true
        type: integer
      - description: update an media
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/swagmodel.InputMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagmodel.Response'
      security:
      - AuthToken: []
      summary: update an media
      tags:
      - media
...
securityDefinitions:
  AuthToken:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"
...